Added script to cleanup duplicates in the publish directory
Updated choco install script to look in the parent and sub-directories for the needed packages

+++ push.install.ps1
@@ -2,4 +2,7 @@ $boxstarterInstallScripts = Join-Path -Resolve $PSSCriptRoot '..\..\BoxStarter'
 $nasInstallScripts = '\\nas\Applications\_BoxStarter'
 
 Copy-Item -Recurse $PSSCriptRoot\Install $boxstarterInstallScripts -Force
-Copy-Item -Recurse $PSSCriptRoot\Install $nasInstallScripts -Force
+
+if (Test-Path $nasInstallScripts) {
+    Copy-Item -Recurse $PSSCriptRoot\Install $nasInstallScripts -Force
+}
